# docs/swagger.yaml
definitions:
...
  admin.UserResponse:
    properties:
      apple_uid:
        type: string
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      invite_by:
        type: integer
      invite_code:
        type: string
      phone:
        type: string
      prefer_signin_method:
        type: string
      realname:
        type: string
      status:
        type: string
      union_id:
        type: string
      updatedAt:
        type: string
      user_type:
        type: string
    type: object
...
  common.Pagination-admin_UserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/admin.UserResponse'
        type: array
      last_page:
        type: integer
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  controllers.WechatPayCreateResponse:
    properties:
      app_id:
        description: APPID APP 支付专用，移动应用AppID
        type: string
      code_url:
        description: CodeURL Web、PC 支付专用，二维码地址
        type: string
      noncestr:
        description: Noncestr APP 支付专用，随机字符串，不长于32位。推荐随机数生成算法
        type: string
      package:
        description: Package APP 支付专用，固定值Sign=WXPay
        type: string
      partner_id:
        description: PartnerID APP 支付专用，商户号mchid对应的值
        type: string
      payment_id:
        type: string
      prepay_id:
        description: PrepayID APP 支付专用，微信返回的支付交易会话ID，该值有效期为2小时。
        type: string
      sandbox:
        type: boolean
      sign:
        description: Sign APP 支付专用，签名，使用字段AppID、timeStamp、nonceStr、prepayid计算得出的签名值
          注意：取值RSA格式
        type: string
      timestamp:
        description: Timestamp APP 支付专用，时间戳 秒级
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: mylxsw@aicode.cc
    name: mylxsw
    url: https://ai.aicode.cc
  description: AIdea API
  title: AIdea API
  version: "1.0"
paths:
...
  /v1/admin/users:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Per Page
        in: query
        name: per_page
        type: integer
      - description: Support searching by phone number, name, and email (prefix fuzzy
          matching)
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Pagination-admin_UserResponse'
      summary: Return the list of all users
      tags:
      - Admin:User
...
  /v1/payment/wechatpay:
    post:
      consumes:
      - application/json
      parameters:
      - description: product id
        in: query
        name: product_id
        required: true
        type: string
      - description: source
        enum:
        - app
        - web
        - pc
        in: query
        name: source
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.WechatPayCreateResponse'
      summary: create wechat payment
      tags:
      - payment
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"